Remove sys submodule from cap-std
Majority of host-specific impls are now done in `cap-primitives`,
and the remaining host-specific bits are few and far between that
they can be included in the main `fs` module in `cap-std`.
