+++ src/fs/dir.rs
@@ -1,9 +1,11 @@
+use crate::fs::{DirBuilder, File, Metadata, OpenOptions, Permissions, ReadDir};
 #[cfg(unix)]
 use crate::os::unix::net::{UnixDatagram, UnixListener, UnixStream};
-use crate::{
-    fs::{DirBuilder, File, Metadata, OpenOptions, Permissions, ReadDir},
-    sys,
-};
+#[cfg(unix)]
+use cap_primitives::fs::symlink;
+use cap_primitives::fs::{canonicalize, link, mkdir, open, stat, unlink, FollowSymlinks};
+#[cfg(windows)]
+use cap_primitives::fs::{symlink_dir, symlink_file};
 #[cfg(unix)]
 use std::os::unix::io::{AsRawFd, FromRawFd, IntoRawFd, RawFd};
 #[cfg(windows)]
@@ -22,22 +24,20 @@ use std::{
 /// Unlike `std::fs`, this API's `canonicalize` returns a relative path since
 /// absolute paths don't interoperate well with the capability model.
 pub struct Dir {
-    sys: sys::fs::Dir,
+    std_file: fs::File,
 }
 
 impl Dir {
     /// Constructs a new instance of `Self` from the given `std::fs::File`.
     #[inline]
     pub fn from_std_file(std_file: fs::File) -> Self {
-        Self {
-            sys: sys::fs::Dir::from_std_file(std_file),
-        }
+        Self { std_file }
     }
 
     /// Consumes `self` and returns a `std::fs::File`.
     #[inline]
     pub fn into_std_file(self) -> fs::File {
-        self.sys.into_std_file()
+        self.std_file
     }
 
     /// Attempts to open a file in read-only mode.
@@ -65,13 +65,26 @@ impl Dir {
         path: P,
         options: &OpenOptions,
     ) -> io::Result<File> {
-        self.sys.open_file_with(path.as_ref(), options)
+        open(&self.std_file, path.as_ref(), options).map(File::from_std)
     }
 
     /// Attempts to open a directory.
     #[inline]
     pub fn open_dir<P: AsRef<Path>>(&self, path: P) -> io::Result<Self> {
-        self.sys.open_dir(path.as_ref())
+        self._open_dir(path.as_ref())
+    }
+
+    #[cfg(unix)]
+    fn _open_dir(&self, path: &Path) -> io::Result<Self> {
+        use std::os::unix::fs::OpenOptionsExt;
+        use yanix::file::OFlag;
+        self.open_file_with(
+            path,
+            OpenOptions::new()
+                .read(true)
+                .custom_flags(OFlag::DIRECTORY.bits()),
+        )
+        .map(|file| crate::fs::Dir::from_std_file(file.std))
     }
 
     /// Creates a new, empty directory at the provided path.
@@ -82,7 +95,7 @@ impl Dir {
     /// [`std::fs::create_dir`]: https://doc.rust-lang.org/std/fs/fn.create_dir.html
     #[inline]
     pub fn create_dir<P: AsRef<Path>>(&self, path: P) -> io::Result<()> {
-        self.sys.create_dir(path.as_ref())
+        mkdir(&self.std_file, path.as_ref())
     }
 
     /// Recursively create a directory and all of its parent components if they are missing.
@@ -146,7 +159,7 @@ impl Dir {
     /// [`std::fs::canonicalize`]: https://doc.rust-lang.org/std/fs/fn.canonicalize.html
     #[inline]
     pub fn canonicalize<P: AsRef<Path>>(&self, path: P) -> io::Result<PathBuf> {
-        self.sys.canonicalize(path.as_ref())
+        canonicalize(&self.std_file, path.as_ref())
     }
 
     /// Copies the contents of one file to another. This function will also copy the permission
@@ -188,7 +201,12 @@ impl Dir {
         dst_dir: &Self,
         dst: Q,
     ) -> io::Result<()> {
-        self.sys.hard_link(src.as_ref(), &dst_dir.sys, dst.as_ref())
+        link(
+            &self.std_file,
+            src.as_ref(),
+            &dst_dir.std_file,
+            dst.as_ref(),
+        )
     }
 
     /// Given a path, query the file system to get information about a file, directory, etc.
@@ -199,7 +217,7 @@ impl Dir {
     /// [`std::fs::metadata`]: https://doc.rust-lang.org/std/fs/fn.metadata.html
     #[inline]
     pub fn metadata<P: AsRef<Path>>(&self, path: P) -> io::Result<Metadata> {
-        self.sys.metadata(path.as_ref())
+        stat(&self.std_file, path.as_ref(), FollowSymlinks::Yes)
     }
 
     /// Returns an iterator over the entries within a directory.
@@ -210,7 +228,11 @@ impl Dir {
     /// [`std::fs::read_dir`]: https://doc.rust-lang.org/std/fs/fn.read_dir.html
     #[inline]
     pub fn read_dir<P: AsRef<Path>>(&self, path: P) -> io::Result<ReadDir> {
-        self.sys.read_dir(path.as_ref())
+        unimplemented!(
+            "Dir::read_dir({:?}, {})",
+            self.std_file,
+            path.as_ref().display()
+        )
     }
 
     /// Read the entire contents of a file into a bytes vector.
@@ -236,7 +258,11 @@ impl Dir {
     /// [`std::fs::read_link`]: https://doc.rust-lang.org/std/fs/fn.read_link.html
     #[inline]
     pub fn read_link<P: AsRef<Path>>(&self, path: P) -> io::Result<PathBuf> {
-        self.sys.read_link(path.as_ref())
+        unimplemented!(
+            "Dir::read_link({:?}, {})",
+            self.std_file,
+            path.as_ref().display()
+        )
     }
 
     /// Read the entire contents of a file into a string.
@@ -261,7 +287,11 @@ impl Dir {
     /// [`std::fs::remove_dir`]: https://doc.rust-lang.org/std/fs/fn.remove_dir.html
     #[inline]
     pub fn remove_dir<P: AsRef<Path>>(&self, path: P) -> io::Result<()> {
-        self.sys.remove_dir(path.as_ref())
+        unimplemented!(
+            "Dir::remove_dir({:?}, {})",
+            self.std_file,
+            path.as_ref().display()
+        )
     }
 
     /// Removes a directory at this path, after removing all its contents. Use carefully!
@@ -272,7 +302,11 @@ impl Dir {
     /// [`std::fs::remove_dir_all`]: https://doc.rust-lang.org/std/fs/fn.remove_dir_all.html
     #[inline]
     pub fn remove_dir_all<P: AsRef<Path>>(&self, path: P) -> io::Result<()> {
-        self.sys.remove_dir_all(path.as_ref())
+        unimplemented!(
+            "Dir::remove_dir_all({:?}, {})",
+            self.std_file,
+            path.as_ref().display()
+        )
     }
 
     /// Removes a file from a filesystem.
@@ -283,7 +317,7 @@ impl Dir {
     /// [`std::fs::remove_file`]: https://doc.rust-lang.org/std/fs/fn.remove_file.html
     #[inline]
     pub fn remove_file<P: AsRef<Path>>(&self, path: P) -> io::Result<()> {
-        self.sys.remove_file(path.as_ref())
+        unlink(&self.std_file, path.as_ref())
     }
 
     /// Rename a file or directory to a new name, replacing the original file if to already exists.
@@ -294,7 +328,12 @@ impl Dir {
     /// [`std::fs::rename`]: https://doc.rust-lang.org/std/fs/fn.rename.html
     #[inline]
     pub fn rename<P: AsRef<Path>, Q: AsRef<Path>>(&self, from: P, to: Q) -> io::Result<()> {
-        self.sys.rename(from.as_ref(), to.as_ref())
+        unimplemented!(
+            "Dir::rename({:?}, {}, {})",
+            self.std_file,
+            from.as_ref().display(),
+            to.as_ref().display()
+        )
     }
 
     /// Changes the permissions found on a file or a directory.
@@ -305,7 +344,12 @@ impl Dir {
     /// [`std::fs::set_permissions`]: https://doc.rust-lang.org/std/fs/fn.set_permissions.html
     #[inline]
     pub fn set_permissions<P: AsRef<Path>>(&self, path: P, perm: Permissions) -> io::Result<()> {
-        self.sys.set_permissions(path.as_ref(), perm)
+        unimplemented!(
+            "Dir::set_permissions({:?}, {}, {:?})",
+            self.std_file,
+            path.as_ref().display(),
+            perm
+        )
     }
 
     /// Query the metadata about a file without following symlinks.
@@ -316,7 +360,7 @@ impl Dir {
     /// [`std::fs::symlink_metadata`]: https://doc.rust-lang.org/std/fs/fn.symlink_metadata.html
     #[inline]
     pub fn symlink_metadata<P: AsRef<Path>>(&self, path: P) -> io::Result<Metadata> {
-        self.sys.symlink_metadata(path.as_ref())
+        stat(&self.std_file, path.as_ref(), FollowSymlinks::No)
     }
 
     /// Write a slice as the entire contents of a file.
@@ -344,10 +388,14 @@ impl Dir {
     #[inline]
     pub fn create_with_dir_builder<P: AsRef<Path>>(
         &self,
-        dir_builder: &DirBuilder,
+        _dir_builder: &DirBuilder,
         path: P,
     ) -> io::Result<()> {
-        self.sys.create_with_dir_builder(dir_builder, path.as_ref())
+        unimplemented!(
+            "Dir::create_with_dir_builder({:?}, {})",
+            self.std_file,
+            path.as_ref().display()
+        )
     }
 
     /// Creates a new symbolic link on a filesystem.
@@ -365,7 +413,7 @@ impl Dir {
     ))]
     #[inline]
     pub fn symlink<P: AsRef<Path>, Q: AsRef<Path>>(&self, src: P, dst: Q) -> io::Result<()> {
-        self.sys.symlink(src.as_ref(), dst.as_ref())
+        symlink(src.as_ref(), &self.std_file, dst.as_ref())
     }
 
     /// Creates a new file symbolic link on a filesystem.
@@ -377,7 +425,7 @@ impl Dir {
     #[cfg(windows)]
     #[inline]
     pub fn symlink_file<P: AsRef<Path>, Q: AsRef<Path>>(&self, src: P, dst: Q) -> io::Result<()> {
-        self.cap_std.symlink_file(src.as_ref(), dst.as_ref())
+        symlink_file(src.as_ref(), &self.std_file, dst.as_ref())
     }
 
     /// Creates a new directory symlink on a filesystem.
@@ -389,7 +437,7 @@ impl Dir {
     #[cfg(windows)]
     #[inline]
     pub fn symlink_dir<P: AsRef<Path>, Q: AsRef<Path>>(&self, src: P, dst: Q) -> io::Result<()> {
-        self.cap_std.symlink_dir(src.as_ref(), dst.as_ref())
+        symlink_dir(src.as_ref(), &self.std_file, dst.as_ref())
     }
 
     /// Creates a new `UnixListener` bound to the specified socket.
@@ -401,7 +449,11 @@ impl Dir {
     #[cfg(unix)]
     #[inline]
     pub fn bind_unix_listener<P: AsRef<Path>>(&self, path: P) -> io::Result<UnixListener> {
-        self.sys.bind_unix_listener(path.as_ref())
+        unimplemented!(
+            "Dir::bind_unix_listener({:?}, {})",
+            self.std_file,
+            path.as_ref().display()
+        )
     }
 
     /// Connects to the socket named by path.
@@ -413,7 +465,11 @@ impl Dir {
     #[cfg(unix)]
     #[inline]
     pub fn connect_unix_stream<P: AsRef<Path>>(&self, path: P) -> io::Result<UnixStream> {
-        self.sys.connect_unix_stream(path.as_ref())
+        unimplemented!(
+            "Dir::connect_unix_stream({:?}, {})",
+            self.std_file,
+            path.as_ref().display()
+        )
     }
 
     /// Creates a Unix datagram socket bound to the given path.
@@ -425,7 +481,11 @@ impl Dir {
     #[cfg(unix)]
     #[inline]
     pub fn bind_unix_datagram<P: AsRef<Path>>(&self, path: P) -> io::Result<UnixDatagram> {
-        self.sys.bind_unix_datagram(path.as_ref())
+        unimplemented!(
+            "Dir::bind_unix_datagram({:?}, {})",
+            self.std_file,
+            path.as_ref().display()
+        )
     }
 
     /// Connects the socket to the specified address.
@@ -438,10 +498,14 @@ impl Dir {
     #[inline]
     pub fn connect_unix_datagram<P: AsRef<Path>>(
         &self,
-        unix_datagram: &UnixDatagram,
+        _unix_datagram: &UnixDatagram,
         path: P,
     ) -> io::Result<()> {
-        self.sys.connect_unix_datagram(unix_datagram, path.as_ref())
+        unimplemented!(
+            "Dir::connect_unix_datagram({:?}, {})",
+            self.std_file,
+            path.as_ref().display()
+        )
     }
 
     /// Sends data on the socket to the specified address.
@@ -454,21 +518,23 @@ impl Dir {
     #[inline]
     pub fn send_to_unix_datagram_addr<P: AsRef<Path>>(
         &self,
-        unix_datagram: &UnixDatagram,
+        _unix_datagram: &UnixDatagram,
         buf: &[u8],
         path: P,
     ) -> io::Result<usize> {
-        self.sys
-            .send_to_unix_datagram_addr(unix_datagram, buf, path.as_ref())
+        unimplemented!(
+            "Dir::send_to_unix_datagram_addr({:?}, {:?}, {})",
+            self.std_file,
+            buf,
+            path.as_ref().display()
+        )
     }
 
     /// Creates a new `Dir` instance that shares the same underlying file handle as the existing
     /// `Dir` instance.
     #[inline]
     pub fn try_clone(&self) -> io::Result<Self> {
-        Ok(Self {
-            sys: self.sys.try_clone()?,
-        })
+        Ok(Self::from_std_file(self.std_file.try_clone()?))
     }
 
     /// Returns `true` if the path points at an existing entity.
@@ -526,7 +592,7 @@ impl FromRawHandle for Dir {
 impl AsRawFd for Dir {
     #[inline]
     fn as_raw_fd(&self) -> RawFd {
-        self.sys.as_raw_fd()
+        self.std_file.as_raw_fd()
     }
 }
 
@@ -534,7 +600,7 @@ impl AsRawFd for Dir {
 impl AsRawHandle for Dir {
     #[inline]
     fn as_raw_handle(&self) -> RawHandle {
-        self.sys.as_raw_handle()
+        self.std_file.as_raw_handle()
     }
 }
 
@@ -542,7 +608,7 @@ impl AsRawHandle for Dir {
 impl IntoRawFd for Dir {
     #[inline]
     fn into_raw_fd(self) -> RawFd {
-        self.sys.into_raw_fd()
+        self.std_file.into_raw_fd()
     }
 }
 
@@ -550,7 +616,7 @@ impl IntoRawFd for Dir {
 impl IntoRawHandle for Dir {
     #[inline]
     fn into_raw_handle(self) -> RawHandle {
-        self.sys.into_raw_handle()
+        self.std_file.into_raw_handle()
     }
 }
 
@@ -567,6 +633,35 @@ fn initial_buffer_size(file: &File) -> usize {
 impl fmt::Debug for Dir {
     // Like libstd's version, but doesn't print the path.
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.sys.fmt(f)
+        let mut b = f.debug_struct("Dir");
+        fmt_debug_dir(&self.std_file, &mut b);
+        b.finish()
     }
+}
+
+#[cfg(any(unix, target_os = "fuchsia"))]
+fn fmt_debug_dir(fd: &impl AsRawFd, b: &mut fmt::DebugStruct) {
+    unsafe fn get_mode(fd: std::os::unix::io::RawFd) -> Option<(bool, bool)> {
+        let mode = yanix::fcntl::get_status_flags(fd);
+        if mode.is_err() {
+            return None;
+        }
+        match mode.unwrap() & yanix::file::OFlag::ACCMODE {
+            yanix::file::OFlag::RDONLY => Some((true, false)),
+            yanix::file::OFlag::RDWR => Some((true, true)),
+            yanix::file::OFlag::WRONLY => Some((false, true)),
+            _ => None,
+        }
+    }
+
+    let fd = fd.as_raw_fd();
+    b.field("fd", &fd);
+    if let Some((read, write)) = unsafe { get_mode(fd) } {
+        b.field("read", &read).field("write", &write);
+    }
+}
+
+#[cfg(windows)]
+fn fmt_debug_dir(fd: &impl AsRawHandle, b: &mut fmt::DebugStruct) {
+    // TODO fill in the blanks
 }
