--- src/sys/winx/mod.rs
@@ -1,1 +1,0 @@
-// TODO: winx implementation
